Linked powers to triggered events
