+++ Buypartisan/Assets/Scripts/Action Scripts.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: cb0f1e42cc156e049a1afe58d793beff
+folderAsset: yes
+timeCreated: 1436911939
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
