+++ Buypartisan/Assets/Scripts/Action Scripts/Action4Script.cs
@@ -12,6 +12,8 @@ public class Action4Script : MonoBehaviour {
 	public GameObject uiController; //this is the UI controller variable. Obtained from the scene
 	private GameObject[] voters; //array which houses the voters. Obtained from the Game Controller
 	private GameObject[] players; //array which houses the players. Obtained from the Game Controller
+	//Brian Mah
+	private RandomEventControllerScript eventController;
 	
 	private int currentPlayer; //this variable finds which player is currently using his turn.
 
@@ -66,6 +68,7 @@ public class Action4Script : MonoBehaviour {
 		if (gameController != null) {
 			voters = gameController.GetComponent<GameController> ().voters;
 			players = gameController.GetComponent<GameController> ().players;
+			eventController = gameController.GetComponent<GameController> ().randomEventController;
 		} else {
 			Debug.Log ("Failed to obtain voters and players array from Game Controller");
 		}
@@ -284,6 +287,9 @@ public class Action4Script : MonoBehaviour {
 		uiController.GetComponent<UI_Script>().toggleActionButtons();
 		this.transform.parent.GetComponent<PlayerTurnsManager> ().IncreaseCostMultiplier();
 		players [currentPlayer].GetComponent<PlayerVariables> ().money -= totalCost;
+		//puts the current player and the event number into the action counter of the event controller
+		//Brian Mah
+		eventController.actionCounter [gameController.GetComponent<GameController>().currentPlayerTurn] [4]++; // the second number should be the number of the action!
 		Destroy(gameObject);
 	}
