+++ Buypartisan/Assets/Scripts/Action Scripts/Action6Script.cs
@@ -1,46 +1,34 @@
-﻿// Chris-chan
-// Michael's standardization notes:
-// PLEASE TALK TO MICHAEL BEFORE WORKING ON ANY ACTIONS
-// Every "1" unit of cost is $10k = 10. Which means no baseCost is below 10.
-// costMultiplier is actually derived from PlayerTurnsManager.cs, which is increased in EndAction(), additively
-// Check if enough money happens in Start(). Subtracting money always happen in EndAction().
-// Please use totalCost for any end calculation, since this will be used to display on the UI's action button
-
+﻿/// <summary>
+/// Sphere of Influence script by Daniel Schlesinger
+/// </summary>
 using UnityEngine;
 using System.Collections;
-/// <summary>
-/// Action0 script by Daniel Schlesinger
-/// 
-/// </summary>
-public class Action0Script : MonoBehaviour {
-    public string actionName = "Voter Suppression";
-	public int baseCost = 100;
-    public int totalCost = 0; // Please use totalCost for any end calculation, since this will be used to display on the UI's action button
-    public float costMultiplier = 1.0f; // Increased by fixed amount within same turn (in PlayerTurnsManager). This is reset to 1 after the END of your turn.
 
+public class Action6Script : MonoBehaviour {
+	public string actionName = "Character Assassination";
+	public int baseCost = 300;
+	public int totalCost = 0; // Please use totalCost for any end calculation, since this will be used to display on the UI's action button
+	public float costMultiplier = 1.0f; // Increased by fixed amount within same turn (in PlayerTurnsManager). This is reset to 1 after the END of your turn.
+	
 	public GameObject gameController; //this is the game controller variable. It is obtained from the scene
 	public GameObject inputManager; //this is the input manager varibale. Obtained from the scene
 	public GameObject uiController; //this is the UI controller variable. Obtained from the scene
 	private GameObject[] voters; //array which houses the voters. Obtained from the Game Controller
 	private GameObject[] players; //array which houses the players. Obtained from the Game Controller
+	//Brian Mah
+	private RandomEventControllerScript eventController;
 
 	private int currentPlayer; //this variable finds which player is currently using his turn.
-	private int selectedVoter = 0;
-	private bool voterSelected = false;
-	private bool foundVoter = false;
-
-	[System.NonSerialized]
-	public bool leftButton = false; //checks if left button has been pressed
-	[System.NonSerialized]
-	public bool rightButton = false; //checks if right button has been pressed
-	[System.NonSerialized]
-	public bool confirmButton = false; //checks if confirm button has been pressed
-	[System.NonSerialized]
-	public bool cancelButton = false;
-
-	//holds the number needed for this action to succeed (Alex Jungroth)
+	private int selectedPlayer;
+	private bool playerSelected = false;
+	private bool foundPlayer = false;
 	public float successRate = 0.5f;
 
+	[System.NonSerialized]
+	public bool confirmButton = false;
+	[System.NonSerialized]
+	public bool cancelButton = false;
+	
 	// Use this for initialization
 	void Start () {
 		gameController = GameObject.FindWithTag ("GameController");
@@ -53,107 +41,69 @@ public class Action0Script : MonoBehaviour {
 		if (gameController != null) {
 			voters = gameController.GetComponent<GameController> ().voters;
 			players = gameController.GetComponent<GameController> ().players;
+			eventController = gameController.GetComponent<GameController> ().randomEventController;
 		} else {
 			Debug.Log ("Failed to obtain voters and players array from Game Controller");
 		}
-
+		
 		//Disables the Action UI buttons
 		uiController.GetComponent<UI_Script>().disableActionButtons();
-
+		
 		//The start function will not end until gets to the end
 		//if you want to destroy the object in the start function,
 		//it has to be the last thing you do, otherwise the flow of
 		//controll will stay with the destroyed instance and 
 		//that will crash the game (Alex Jungroth)
-
+		
 		//Get's whose turn it is from the gameController. Then checks if he has enough money to perform the action
 		currentPlayer = gameController.GetComponent<GameController> ().currentPlayerTurn;
 		costMultiplier = this.transform.parent.GetComponent<PlayerTurnsManager> ().costMultiplier;
-
-		this.transform.position = voters [selectedVoter].transform.position;
-
 		if (players[currentPlayer].GetComponent<PlayerVariables> ().money < (baseCost * costMultiplier)) {
 			Debug.Log ("Current Player doesn't have enough money to make this action.");
-
+			
 			uiController.GetComponent<UI_Script>().toggleActionButtons();
 			Destroy(gameObject);
 		}
-        else
-        {
-            totalCost = (int)(baseCost * costMultiplier);
-        }
+		else
+		{
+			totalCost = (int)(baseCost * costMultiplier);
+		}
 	}
-	
+
 	// Update is called once per frame
 	void Update () {
-
+		//Debug.Log ("here");
 		//ends the action if the cancel button is pressed (Alex Jungroth)
 		if (cancelButton) {
-			//handles early canceling(Alex Jungroth)
 			uiController.GetComponent<UI_Script> ().activateAction0UI2 ();
 			uiController.GetComponent<UI_Script> ().toggleActionButtons ();
 			Destroy (gameObject);
 		}
-
-		if (!voterSelected) {
-			/*if (leftButton) {
-				if (selectedVoter == 0) {
-					selectedVoter = voters.Length - 1;
-				} else {
-					selectedVoter -= 1;
-				}
-				this.transform.position = voters [selectedVoter].transform.position;
-				
-				leftButton = false;
-			}
-			if (rightButton) {
-				if (selectedVoter == (voters.Length - 1)) {
-					selectedVoter = 0;
-				} else {
-					selectedVoter += 1;
-				}
-				this.transform.position = voters [selectedVoter].transform.position;
-
-				rightButton = false;
-			}
-			if (confirmButton) {
-				Debug.Log ("Here");
-				voterSelected = true;
-				confirmButton = false;
-				uiController.GetComponent<UI_Script> ().activateAction0UI2 ();
-
-				//checks to see if the power succeeded (Alex Jungroth)
-
-			}*/
-
-			if (Input.GetMouseButtonDown(0)) {
-				for (int i = 0; i < voters.Length; i++) {
-					if (voters [i].GetComponent<VoterVariables> ().GetSelected ()) {
-							
-						if (Random.value >= successRate)
-						{
-							voters [i].GetComponent<VoterVariables> ().votes = 0;
-						}
-						foundVoter = true;
-						voterSelected = true;
+		if (Input.GetMouseButtonDown (0)) {
+			for (selectedPlayer = 0; selectedPlayer < players.Length; selectedPlayer++) {
+				if(players[selectedPlayer].GetComponent<PlayerVariables>().GetSelected()) {
+					foundPlayer = true;
+					playerSelected = true;
+					Debug.Log (playerSelected);
+					if (Random.value >= successRate) {
+						players [selectedPlayer].GetComponent<PlayerVariables> ().sphereController.transform.localScale -= new Vector3 (10f, 10f, 10f);
 					}
 				}
-				if (!foundVoter) {
-					Debug.Log ("Please Select A Voter by Mousing Over Them and Clicking");
-				}
-
 			}
 		}
-		if (voterSelected)
+		if (playerSelected) {
 			EndAction ();
-
-	
+		}
 	}
+	
 	void EndAction() {
-		uiController.GetComponent<UI_Script>().activateAction0UI2();
 		uiController.GetComponent<UI_Script>().toggleActionButtons();
 		this.transform.parent.GetComponent<PlayerTurnsManager> ().IncreaseCostMultiplier();
 		players [currentPlayer].GetComponent<PlayerVariables> ().money -= totalCost;  // Money is subtracted
+		//puts the current player and the event number into the action counter of the event controller
+		//Brian Mah
+		eventController.actionCounter [gameController.GetComponent<GameController>().currentPlayerTurn] [6]++; // the second number should be the number of the action!
 		Destroy(gameObject);
 	}
 }
+
